requirements.txt
gpiozero
PIL
luma.core
